# LMS_CODEBASE/packages/openapi/ulms-api.yaml
openapi: 3.0.3
info:
  title: ULMS API
  version: 1.7.0
  description: |
    Unisoft Loan Management System — public API (source of truth per PLANNING/05).
    Conventions: problem+json errors (stable `code`, registry docs/errors.md),
    Idempotency-Key on creating POSTs (replays return the original response
    with `Idempotent-Replay: true`, 48h window), list responses use the
    {data, meta} envelope, money is integer minor units, bilingual objects
    {en,bn} render ONE per request.
    Mutating routes are role-gated (403 ULMS-FORBIDDEN) — realm roles per 06 §1;
    the approval act() ladder-level check maps roles per the band table.
    v1.2 (P2): assessments (CIB/score/DBR/collateral), disbursements with dual
    authorization, compliance EOD/classification/calculator.
    v1.2.1: CIB file channel + 24-month repayment track + raw kept only in the
    encrypted object store (06 §5) + approver inbox.
    v1.3: optimistic concurrency (PATCH version ⇒ 409 ULMS-STATE-0002),
    collateral valuation history, application detail incl. workflow state,
    ?officer= list filter, CIB retry×5 + RB-05 alert on outage.
    v1.4 (P3): payments + rails (intent→redirect→HMAC webhook idempotent by
    UTR), schedules/statements (JSON+CSV), settle quotes, reschedules,
    collections (worklist/actions/PTP/field tasks), tax certificates,
    borrower portal projections.
    v1.4.1: ESCALATE approval action, dunning-ladder timers (queue/run/
...
    portfolio reporting read model.
...
servers:
...
security: [{bearerAuth: []}]
paths:
...
  /assessments/cib/file:
    post:
      tags: [assessments]
      summary: Monthly SFTP file channel (11 §1 parseFile) — same parser as real-time; deduped by (cif, period, file); raw stored encrypted (06 §5)
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [cif, period, fileId, raw], properties: {cif: {type: string}, period: {type: string, description: YYYY-MM}, fileId: {type: string}, raw: {type: string, description: fixed-width file body}}}
      responses:
        "200": {description: pull status, content: {application/json: {schema: {$ref: "#/components/schemas/CibPull"}}}}
        "404": {$ref: "#/components/responses/Problem"}
  /assessments/cib/{cif}:
    post:
      tags: [assessments]
      summary: CIB pull (11 §1 — mock real-time in P2; dedupe by cif/period/file; malformed rows quarantined)
      parameters: [{name: cif, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: pull status, content: {application/json: {schema: {$ref: "#/components/schemas/CibPull"}}}}
    get:
      tags: [assessments]
      summary: Latest parsed report + facilities (CIB viewer)
      parameters: [{name: cif, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: detail, content: {application/json: {schema: {$ref: "#/components/schemas/CibDetail"}}}}
        "404": {$ref: "#/components/responses/Problem"}
  /assessments/{applicationId}/score:
    post:
      tags: [assessments]
      summary: (Re)run assessment — server-computed DBR (income + bureau installments + EMI oracle) + versioned scorecard
      parameters: [{name: applicationId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AssessmentFacts"}
      responses:
        "200": {description: outcome, content: {application/json: {schema: {$ref: "#/components/schemas/Score"}}}}
        "422": {$ref: "#/components/responses/Problem"}
    get:
      tags: [assessments]
      summary: Latest stored score
      parameters: [{name: applicationId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        "200": {description: score, content: {application/json: {schema: {$ref: "#/components/schemas/Score"}}}}
        "404": {$ref: "#/components/responses/Problem"}
  /assessments/{applicationId}/dbr:
    post:
      tags: [assessments]
      summary: DBR-only computation with the policy breakdown (income + bureau installments + EMI oracle)
      parameters: [{name: applicationId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AssessmentFacts"}
      responses:
        "200": {description: breakdown, content: {application/json: {schema: {$ref: "#/components/schemas/DbrBreakdown"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /assessments/{applicationId}/collateral:
    post:
      tags: [assessments]
      summary: Add collateral (valuation + insurance tracking)
      parameters: [{name: applicationId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [type, valueMinor], properties: {type: {type: string, enum: [LAND, PROPERTY, FDR, MACHINERY, GUARANTEE]}, description: {type: string}, valueMinor: {type: integer}, valuedOn: {type: string, format: date}, insuredUntil: {type: string, format: date}}}
      responses:
        "200": {description: created, content: {application/json: {schema: {$ref: "#/components/schemas/Collateral"}}}}
    get:
      tags: [assessments]
      summary: List collateral for an application
      parameters: [{name: applicationId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/Collateral"}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
...
  /assessments/collateral/{collateralId}/valuation:
    post:
      tags: [assessments]
      summary: Revalue collateral — appends an immutable history row + updates current (04 §3)
      parameters: [{name: collateralId, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [valueMinor], properties: {valueMinor: {type: integer}, valuedOn: {type: string, format: date}}}
      responses:
        "200": {description: updated collateral, content: {application/json: {schema: {$ref: "#/components/schemas/Collateral"}}}}
        "404": {$ref: "#/components/responses/Problem"}
    get:
      tags: [assessments]
      summary: Full valuation history (04 §3 collateral_valuation)
      parameters: [{name: collateralId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        "200": {description: list envelope, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {type: object, properties: {id: {type: string, format: uuid}, valueMinor: {type: integer}, valuedOn: {type: string, nullable: true}, valuedBy: {type: string}, createdAt: {type: string}}}}, meta: {$ref: "#/components/schemas/Meta"}}}}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  responses:
    Problem:
      description: error
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        code: {type: string, description: "stable code — registry docs/errors.md (ULMS-NOT-FOUND, ULMS-FORBIDDEN, ULMS-VAL-0001, ULMS-STATE-0001)"}
        detail: {type: string}
        requestId: {type: string, format: uuid}
        fields:
          type: array
          items:
            type: object
            properties: {path: {type: string}, code: {type: string}, message: {type: string}, messageBn: {type: string}}
    Meta:
      type: object
      properties:
        requestId: {type: string, format: uuid}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer}
        totalPages: {type: integer}
...
    CibPull:
      type: object
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [PENDING, PARSED, FAILED]}
        period: {type: string, description: YYYY-MM}
        error: {type: string, nullable: true}
    CibDetail:
      type: object
      properties:
        id: {type: string, format: uuid}
        status: {type: string}
        period: {type: string}
        pulledAt: {type: string, format: date-time}
        facilities:
          type: array
          items:
            type: object
            properties:
              lenderCode: {type: string}
              lenderName: {type: string}
              facilityType: {type: string}
              limitMinor: {type: integer}
              outstandingMinor: {type: integer}
              overdueMinor: {type: integer}
              installmentMinor: {type: integer}
              dpd: {type: integer}
              classification: {type: string}
              lastPaymentDate: {type: string, nullable: true}
              repaymentTrack: {type: string, description: "24 digits — '0' clean · '1' overdue · '2' missed, oldest→latest"}
        pullHistory:
          type: array
          description: every pull for the subject (03 "parsed facilities + history")
          items:
            type: object
            properties: {period: {type: string}, status: {type: string}, source: {type: string}, pulledAt: {type: string}}
    AssessmentFacts:
      type: object
      required: [customerId, amountMinor, tenorMonths, rateType]
      properties:
        customerId: {type: string, format: uuid}
        amountMinor: {type: integer}
        tenorMonths: {type: integer}
        rateType: {type: string, enum: [FIXED, FLOATING]}
        incomeMinor: {type: integer, nullable: true}
        existingEmiMinor: {type: integer, nullable: true}
        collateralValueMinor: {type: integer, nullable: true}
    Score:
      type: object
      properties:
        score: {type: integer, description: 300–900}
        grade: {type: string, enum: [A, B, C, D]}
        decision: {type: string, enum: [AUTO_PASS, REFER, AUTO_DECLINE]}
        version: {type: integer, description: scorecard policy version (03 score versioning)}
        dbrPercent: {type: string, nullable: true}
        cibObligationMinor: {type: integer}
        proposedEmiMinor: {type: integer}
        factors: {type: string, description: JSON factor→points breakdown}
    DbrBreakdown:
      type: object
      properties:
        dbrPercent: {type: string, nullable: true, description: null when income ≤ 0}
        incomeMinor: {type: integer}
        existingEmiMinor: {type: integer}
        cibObligationMinor: {type: integer}
        proposedEmiMinor: {type: integer}
    Collateral:
      type: object
      properties:
        id: {type: string, format: uuid}
        type: {type: string}
        description: {type: string}
        valueMinor: {type: integer}
        valuedOn: {type: string, nullable: true}
        insuredUntil: {type: string, nullable: true}
...